html string builder thing
Hey, not sure if any of this is useful for the lab. I was working on something that would read in an html file and build a string from it, once it reads "<!StartHere>" we can add in new values. We would have to have a method in that if(str.equals()){...} that can add as many employee strings as we wanted and appending them to the string.